Add validation auth token & locking mutations: Users, Posts, Comments

+++ src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
+import getUserId from '../utils/getUserId';
 
 const Mutation = {
   async createUser(parent, args, { prisma }, info) {
@@ -37,21 +38,24 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, 'secretkey'),
     };
   },
-  deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({ where: { id: args.id }, info });
+  deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    return prisma.mutation.deleteUser({ where: { id: userId }, info });
   },
-  updateUser(parent, args, { prisma }, info) {
+  updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
       info,
     );
   },
-  createPost(parent, args, { prisma }, info) {
+  createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createPost(
       {
         data: {
@@ -60,7 +64,7 @@ const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
         },
@@ -68,10 +72,30 @@ const Mutation = {
       info,
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
     return prisma.mutation.deletePost({ where: { id: args.id } }, info);
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error('Unable to update post');
+    }
     return prisma.mutation.updatePost(
       {
         where: {
@@ -82,14 +106,15 @@ const Mutation = {
       info,
     );
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
           post: {
@@ -102,7 +127,17 @@ const Mutation = {
       info,
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!commentExists) {
+      throw new Error('Unable to delete comment');
+    }
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -112,16 +147,23 @@ const Mutation = {
       info,
     );
   },
-  updateComment(parent, args, { prisma }, info) {
-    return prisma.mutation.updateComment(
-      {
-        where: {
-          id: args.id,
-        },
-        data: args.data,
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
       },
-      info,
-    );
+    });
+    if (!commentExists) {
+      throw new Error('Unable to update comment');
+    }
+    return prisma.mutation.updateComment({
+      where: {
+        id: args.id,
+      },
+      data: args.data,
+    }, info);
   },
 };
